==> src/08-termos-controversos-valicadao.r <==
"Termos não detectados"
# $`1`
# # A tibble: 4 × 3
#       n termo                            detectado
#   <int> <chr>                            <lgl>    
# 1 16890 menciona_lai_resposta            TRUE     
# 2 16890 menciona_sigilo_resposta         FALSE    
# 3 16890 menciona_dados_pessoais_resposta FALSE    
# 4 16890 menciona_lgpd_resposta           FALSE    

"Termos não detectados"
# $`2`
# # A tibble: 4 × 3
#       n termo                            detectado
#   <int> <chr>                            <lgl>    
# 1 14185 menciona_lai_resposta            FALSE    
# 2 14185 menciona_sigilo_resposta         FALSE    
# 3 14185 menciona_dados_pessoais_resposta FALSE    
# 4 14185 menciona_lgpd_resposta           FALSE    

"Menciona somente termo sigilo"
# $`3`
# # A tibble: 4 × 3
#       n termo                            detectado
#   <int> <chr>                            <lgl>    
# 1  5167 menciona_lai_resposta            TRUE     
# 2  5167 menciona_sigilo_resposta         TRUE     
# 3  5167 menciona_dados_pessoais_resposta FALSE    
# 4  5167 menciona_lgpd_resposta           FALSE    

"Menciona os termos sigilo e dados pessoais"
# $`4`
# # A tibble: 4 × 3
#       n termo                            detectado
#   <int> <chr>                            <lgl>    
# 1  2693 menciona_lai_resposta            TRUE     
# 2  2693 menciona_sigilo_resposta         TRUE     
# 3  2693 menciona_dados_pessoais_resposta TRUE     
# 4  2693 menciona_lgpd_resposta           FALSE    

"Menciona somente dados pessoais"
# $`5`
# # A tibble: 4 × 3
#       n termo                            detectado
#   <int> <chr>                            <lgl>    
# 1  1683 menciona_lai_resposta            TRUE     
# 2  1683 menciona_sigilo_resposta         FALSE    
# 3  1683 menciona_dados_pessoais_resposta TRUE     
# 4  1683 menciona_lgpd_resposta           FALSE    

"Menciona somente sigilo"
# $`6`
# # A tibble: 4 × 3
#       n termo                            detectado
#   <int> <chr>                            <lgl>    
# 1  1647 menciona_lai_resposta            FALSE    
# 2  1647 menciona_sigilo_resposta         TRUE     
# 3  1647 menciona_dados_pessoais_resposta FALSE    
# 4  1647 menciona_lgpd_resposta           FALSE    

"Menciona somente dados pessoais"
# $`7`
# # A tibble: 4 × 3
#       n termo                            detectado
#   <int> <chr>                            <lgl>    
# 1   396 menciona_lai_resposta            FALSE    
# 2   396 menciona_sigilo_resposta         FALSE    
# 3   396 menciona_dados_pessoais_resposta TRUE     
# 4   396 menciona_lgpd_resposta           FALSE    

"Menciona os termos sigilo e dados pessoais"
# $`8`
# # A tibble: 4 × 3
#       n termo                            detectado
#   <int> <chr>                            <lgl>    
# 1   305 menciona_lai_resposta            FALSE    
# 2   305 menciona_sigilo_resposta         TRUE     
# 3   305 menciona_dados_pessoais_resposta TRUE     
# 4   305 menciona_lgpd_resposta           FALSE    

"Menciona os termos dados pessoais e LGPD"
# $`9`
# # A tibble: 4 × 3
#       n termo                            detectado
#   <int> <chr>                            <lgl>    
# 1   202 menciona_lai_resposta            TRUE     
# 2   202 menciona_sigilo_resposta         FALSE    
# 3   202 menciona_dados_pessoais_resposta TRUE     
# 4   202 menciona_lgpd_resposta           TRUE     

"Menciona os termos sigilo, dados pessoais e LGPD"
# $`10`
# # A tibble: 4 × 3
#       n termo                            detectado
#   <int> <chr>                            <lgl>    
# 1   186 menciona_lai_resposta            TRUE     
# 2   186 menciona_sigilo_resposta         TRUE     
# 3   186 menciona_dados_pessoais_resposta TRUE     
# 4   186 menciona_lgpd_resposta           TRUE     

 "Menciona os termos dados pessoais e LGPD"
# $`11`
# # A tibble: 4 × 3
#       n termo                            detectado
#   <int> <chr>                            <lgl>    
# 1    54 menciona_lai_resposta            FALSE    
# 2    54 menciona_sigilo_resposta         FALSE    
# 3    54 menciona_dados_pessoais_resposta TRUE     
# 4    54 menciona_lgpd_resposta           TRUE     

"Menciona somente LGPD"
# $`12`
# # A tibble: 4 × 3
#       n termo                            detectado
#   <int> <chr>                            <lgl>    
# 1    44 menciona_lai_resposta            TRUE     
# 2    44 menciona_sigilo_resposta         FALSE    
# 3    44 menciona_dados_pessoais_resposta FALSE    
# 4    44 menciona_lgpd_resposta           TRUE     

"Menciona somente LGPD"
# $`13`
# # A tibble: 4 × 3
#       n termo                            detectado
#   <int> <chr>                            <lgl>    
# 1    32 menciona_lai_resposta            FALSE    
# 2    32 menciona_sigilo_resposta         FALSE    
# 3    32 menciona_dados_pessoais_resposta FALSE    
# 4    32 menciona_lgpd_resposta           TRUE     

"Menciona os termos LGPD e sigilo"
# $`14`
# # A tibble: 4 × 3
#       n termo                            detectado
#   <int> <chr>                            <lgl>    
# 1    15 menciona_lai_resposta            TRUE     
# 2    15 menciona_sigilo_resposta         TRUE     
# 3    15 menciona_dados_pessoais_resposta FALSE    
# 4    15 menciona_lgpd_resposta           TRUE     

"Menciona os termos sigilo, dados pessoais e LGPD"
# $`15`
# # A tibble: 4 × 3
#       n termo                            detectado
#   <int> <chr>                            <lgl>    
# 1    14 menciona_lai_resposta            FALSE    
# 2    14 menciona_sigilo_resposta         TRUE     
# 3    14 menciona_dados_pessoais_resposta TRUE     
# 4    14 menciona_lgpd_resposta           TRUE     

"Menciona os termos sigilo e LGPD"
# $`16`
# # A tibble: 4 × 3
#       n termo                            detectado
#   <int> <chr>                            <lgl>    
# 1     4 menciona_lai_resposta            FALSE    
# 2     4 menciona_sigilo_resposta         TRUE     
# 3     4 menciona_dados_pessoais_resposta FALSE    
# 4     4 menciona_lgpd_resposta           TRUE
